moved evaluation methods to separate file
